+++ NEWS.md
@@ -1,8 +1,9 @@
-# connector 0.1.1.9002 (development version)
+# connector 0.1.1.9*** (development version)
 
 ## Enhancements
 * Added upload_cnt and download_cnt methods for ConnectorLogger
 * Added resource validation system with `validate_resource()` function and `check_resource()` S3 methods
+* Added "metadata" attribute to connectors object
 
 ## Bugs